Move validateCertificateTrust() into Certificate class
